refactor(registry): move auth tests behind facade
