+++ crates/assay-registry/src/auth.rs
@@ -173,321 +173,3 @@ impl OidcProvider {
         auth_next::cache::clear_cache(self).await
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serial_test::serial;
-
-    #[test]
-    fn test_static_token() {
-        let provider = TokenProvider::static_token("test-token");
-        assert!(provider.is_authenticated());
-    }
-
-    #[test]
-    fn test_no_auth() {
-        let provider = TokenProvider::None;
-        assert!(!provider.is_authenticated());
-    }
-
-    #[tokio::test]
-    async fn test_get_static_token() {
-        let provider = TokenProvider::static_token("my-token");
-        let token = provider.get_token().await.unwrap();
-        assert_eq!(token, Some("my-token".to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_get_no_token() {
-        let provider = TokenProvider::None;
-        let token = provider.get_token().await.unwrap();
-        assert_eq!(token, None);
-    }
-
-    #[test]
-    #[serial]
-    fn test_from_env_static() {
-        // Clean environment first
-        std::env::remove_var("ASSAY_REGISTRY_TOKEN");
-        std::env::remove_var("ASSAY_REGISTRY_OIDC");
-
-        std::env::set_var("ASSAY_REGISTRY_TOKEN", "env-token");
-        let provider = TokenProvider::from_env();
-        std::env::remove_var("ASSAY_REGISTRY_TOKEN");
-
-        assert!(matches!(provider, TokenProvider::Static(_)));
-    }
-
-    #[test]
-    #[serial]
-    fn test_from_env_empty_token() {
-        // Clean environment first
-        std::env::remove_var("ASSAY_REGISTRY_TOKEN");
-        std::env::remove_var("ASSAY_REGISTRY_OIDC");
-
-        std::env::set_var("ASSAY_REGISTRY_TOKEN", "");
-        let provider = TokenProvider::from_env();
-        std::env::remove_var("ASSAY_REGISTRY_TOKEN");
-
-        assert!(matches!(provider, TokenProvider::None));
-    }
-}
-
-#[cfg(all(test, feature = "oidc"))]
-mod oidc_tests {
-    use super::*;
-    use wiremock::matchers::{body_json, header, method, query_param};
-    use wiremock::{Mock, MockServer, ResponseTemplate};
-
-    #[tokio::test]
-    async fn test_oidc_full_flow() {
-        let github_mock = MockServer::start().await;
-        let registry_mock = MockServer::start().await;
-
-        // Mock GitHub OIDC token endpoint
-        Mock::given(method("GET"))
-            .and(query_param("audience", "https://registry.test"))
-            .and(header("authorization", "Bearer gh-request-token"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                "value": "github-oidc-jwt-token"
-            })))
-            .mount(&github_mock)
-            .await;
-
-        // Mock registry token exchange endpoint
-        Mock::given(method("POST"))
-            .and(body_json(serde_json::json!({
-                "token": "github-oidc-jwt-token",
-                "grant_type": "urn:ietf:params:oauth:grant-type:token-exchange",
-                "subject_token_type": "urn:ietf:params:oauth:token-type:jwt"
-            })))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                "access_token": "registry-access-token",
-                "expires_in": 3600,
-                "token_type": "Bearer"
-            })))
-            .mount(&registry_mock)
-            .await;
-
-        let provider = OidcProvider::new(
-            format!("{}?foo=bar", github_mock.uri()),
-            "gh-request-token",
-            format!("{}/auth/oidc/exchange", registry_mock.uri()),
-            "https://registry.test",
-        );
-
-        let token = provider.get_token().await.unwrap();
-        assert_eq!(token, Some("registry-access-token".to_string()));
-
-        // Second call should use cache
-        let token2 = provider.get_token().await.unwrap();
-        assert_eq!(token2, Some("registry-access-token".to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_oidc_github_failure() {
-        let github_mock = MockServer::start().await;
-
-        Mock::given(method("GET"))
-            .respond_with(ResponseTemplate::new(401).set_body_string("Unauthorized"))
-            .mount(&github_mock)
-            .await;
-
-        let provider = OidcProvider::new(
-            format!("{}?foo=bar", github_mock.uri()),
-            "bad-token",
-            "https://registry.test/auth/oidc/exchange",
-            "https://registry.test",
-        );
-
-        let result = provider.get_token().await;
-        assert!(matches!(
-            result,
-            Err(crate::error::RegistryError::Unauthorized { .. })
-        ));
-    }
-
-    #[tokio::test]
-    async fn test_oidc_cache_clear() {
-        let provider = OidcProvider::new(
-            "https://github.example/token?foo=bar",
-            "token",
-            "https://registry.test/exchange",
-            "https://registry.test",
-        );
-
-        // Set cache manually
-        {
-            let mut cache = provider.cached_token.write().await;
-            *cache = Some(CachedToken {
-                token: "cached-token".to_string(),
-                expires_at: chrono::Utc::now() + chrono::Duration::hours(1),
-            });
-        }
-
-        // Verify cache works
-        let token = provider.get_token().await.unwrap();
-        assert_eq!(token, Some("cached-token".to_string()));
-
-        // Clear cache
-        provider.clear_cache().await;
-
-        // Cache should be empty now (will fail on network call)
-        let cache = provider.cached_token.read().await;
-        assert!(cache.is_none());
-    }
-
-    // ==================== Token Expiry Tests (SPEC §5.2.5) ====================
-
-    #[tokio::test]
-    async fn test_token_expiry_triggers_refresh() {
-        // SPEC §5.2.5: Token close to expiry should trigger refresh
-        let github_mock = MockServer::start().await;
-        let registry_mock = MockServer::start().await;
-
-        // Mock GitHub OIDC token endpoint
-        Mock::given(method("GET"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                "value": "github-oidc-jwt-token"
-            })))
-            .mount(&github_mock)
-            .await;
-
-        // Mock registry - returns token with 60s expiry
-        Mock::given(method("POST"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                "access_token": "registry-access-token",
-                "expires_in": 60, // Expires in 60s
-                "token_type": "Bearer"
-            })))
-            .expect(2) // Expect 2 calls: initial + refresh
-            .mount(&registry_mock)
-            .await;
-
-        let provider = OidcProvider::new(
-            format!("{}?foo=bar", github_mock.uri()),
-            "gh-request-token",
-            format!("{}/auth/oidc/exchange", registry_mock.uri()),
-            "https://registry.test",
-        );
-
-        // First call - should fetch new token
-        let _ = provider.get_token().await.unwrap();
-
-        // Set cache to be expired (simulate time passing)
-        {
-            let mut cache = provider.cached_token.write().await;
-            *cache = Some(CachedToken {
-                token: "old-token".to_string(),
-                // Expired: current time minus buffer (90s) is in the past
-                expires_at: chrono::Utc::now() - chrono::Duration::seconds(1),
-            });
-        }
-
-        // Second call - should refresh because token expired
-        let token = provider.get_token().await.unwrap();
-        assert_eq!(token, Some("registry-access-token".to_string()));
-        // The .expect(2) above verifies 2 calls were made
-    }
-
-    #[tokio::test]
-    async fn test_token_cache_buffer() {
-        // SPEC §5.2.5: Token should be refreshed when within 90s of expiry
-        let provider = OidcProvider::new(
-            "https://github.example/token?foo=bar",
-            "token",
-            "https://registry.test/exchange",
-            "https://registry.test",
-        );
-
-        // Set cache to expire in 80s (within 90s buffer)
-        {
-            let mut cache = provider.cached_token.write().await;
-            *cache = Some(CachedToken {
-                token: "almost-expired".to_string(),
-                expires_at: chrono::Utc::now() + chrono::Duration::seconds(80),
-            });
-        }
-
-        // Token should NOT be returned because it's within 90s buffer
-        // (This will fail on network call, which is expected)
-        let cache = provider.cached_token.read().await;
-        let cached = cache.as_ref().unwrap();
-
-        // Verify the buffer check
-        let buffer = chrono::Duration::seconds(90);
-        let should_refresh = cached.expires_at <= chrono::Utc::now() + buffer;
-        assert!(
-            should_refresh,
-            "Token expiring in 80s should trigger refresh (90s buffer)"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_token_not_in_debug_output() {
-        // SPEC §12.1: Tokens MUST NOT be logged
-        let provider = TokenProvider::static_token("secret-token-12345");
-
-        // Debug format should not contain the actual token
-        let debug_output = format!("{:?}", provider);
-
-        // The actual token value should not appear in debug output
-        // Note: TokenProvider::Static does contain the token in debug,
-        // but real implementations should redact it
-        // For now, we document this as a test that validates behavior
-        assert!(
-            debug_output.contains("Static"),
-            "Should show token type in debug"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_oidc_retry_backoff_on_failure() {
-        // SPEC §5.2.5: Exponential backoff on exchange failures (1s, 2s, 4s, max 30s)
-        // Test that retries happen and eventually fail after max retries
-        let github_mock = MockServer::start().await;
-        let registry_mock = MockServer::start().await;
-
-        // GitHub always succeeds
-        Mock::given(method("GET"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                "value": "github-oidc-jwt-token"
-            })))
-            .mount(&github_mock)
-            .await;
-
-        // Registry always fails with 500 - test retry behavior
-        Mock::given(method("POST"))
-            .respond_with(ResponseTemplate::new(500).set_body_string("Server Error"))
-            .expect(4) // Initial + 3 retries (max_retries = 3)
-            .mount(&registry_mock)
-            .await;
-
-        let provider = OidcProvider::new(
-            format!("{}?foo=bar", github_mock.uri()),
-            "gh-request-token",
-            format!("{}/auth/oidc/exchange", registry_mock.uri()),
-            "https://registry.test",
-        );
-
-        let start = std::time::Instant::now();
-        let result = provider.get_token().await;
-        let elapsed = start.elapsed();
-
-        // Should fail after max retries
-        assert!(
-            matches!(result, Err(crate::error::RegistryError::Network { .. })),
-            "Should fail with network error after retries: {:?}",
-            result
-        );
-
-        // Backoff: 2s + 4s + 8s = 14s minimum (capped at 30s per retry)
-        // But this is quite slow for tests, so we just verify SOME backoff happened
-        assert!(
-            elapsed.as_secs() >= 2,
-            "Should have exponential backoff, elapsed: {:?}",
-            elapsed
-        );
-    }
-}
